Create world-clock-extension.js
This extension class defines a single custom block, getTime, which is a reporter block that returns the current time from the world clock at https://greenwichmeantime.com/timepiece/world-clock/ as a string. The getTime function uses the JavaScript fetch API to make a request to https://greenwichmeantime.com/timepiece/world-clock/, gets the response as text, and uses the DOMParser API to parse the HTML and extract the current time from the page. The function returns a promise that resolves with the current time string or rejects with an error if there was a problem with the request or parsing the HTML.

The constructor function is used to initialize the extension and can be used to define any properties or perform any setup tasks that are required by the extension. In this example, the constructor function is empty because it is not needed for this extension.